[core] Move platform helper implementations into their own file
